Refactor: Move send_openmrs_data
